Rename base to goflux
