+++ .goreleaser.yml
@@ -94,20 +94,24 @@ release:
     Download the appropriate binary for your platform from the assets below.
 
 # Homebrew tap
-brews:
+homebrew_casks:
   - name: flux
+    binary: flux
     repository:
       owner: barisgit
       name: homebrew-goflux
       token: "{{ .Env.HOMEBREW_TAP_TOKEN }}"
     directory: Formula
+    commit_author:
+      name: goreleaserbot
+      email: bot@goreleaser.com
+    commit_msg_template: "Brew cask update for {{ .ProjectName }} version {{ .Tag }}"
     homepage: https://github.com/barisgit/goflux
     description: "🚀 The fastest way to build full-stack applications with Go + TypeScript"
     license: MIT
-    install: |
-      bin.install "flux"
-    test: |
-      system "#{bin}/flux", "--version"
+    custom_block: |
+      head "https://github.com/barisgit/goflux.git"
+    caveats: "Run 'flux --help' to see available commands"
 
 # Chocolatey package for Windows (commented out for cross-platform development)
 # chocolateys:
