+++ adapters/echo/static.go
@@ -3,14 +3,14 @@ package echo
 import (
 	"embed"
 
-	"github.com/barisgit/goflux/base"
+	"github.com/barisgit/goflux/goflux"
 	"github.com/labstack/echo/v4"
 )
 
 // StaticHandler creates an Echo handler using the shared static logic
-func StaticHandler(assets embed.FS, config base.StaticConfig) echo.HandlerFunc {
+func StaticHandler(assets embed.FS, config goflux.StaticConfig) echo.HandlerFunc {
 	return func(c echo.Context) error {
-		response := base.ServeStaticFile(assets, config, c.Request().URL.Path)
+		response := goflux.ServeStaticFile(assets, config, c.Request().URL.Path)
 
 		if response.NotFound {
 			return c.NoContent(404)
