+++ adapters/fiber/static.go
@@ -3,14 +3,14 @@ package fiber
 import (
 	"embed"
 
-	"github.com/barisgit/goflux/base"
+	"github.com/barisgit/goflux/goflux"
 	"github.com/gofiber/fiber/v2"
 )
 
 // StaticHandler creates a Fiber handler using the shared static logic
-func StaticHandler(assets embed.FS, config base.StaticConfig) fiber.Handler {
+func StaticHandler(assets embed.FS, config goflux.StaticConfig) fiber.Handler {
 	return func(c *fiber.Ctx) error {
-		response := base.ServeStaticFile(assets, config, c.Path())
+		response := goflux.ServeStaticFile(assets, config, c.Path())
 
 		if response.NotFound {
 			return c.SendStatus(404)
