+++ adapters/gin/static.go
@@ -3,14 +3,14 @@ package gin
 import (
 	"embed"
 
-	"github.com/barisgit/goflux/base"
+	"github.com/barisgit/goflux/goflux"
 	"github.com/gin-gonic/gin"
 )
 
 // StaticHandler creates a Gin handler using the shared static logic
-func StaticHandler(assets embed.FS, config base.StaticConfig) gin.HandlerFunc {
+func StaticHandler(assets embed.FS, config goflux.StaticConfig) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		response := base.ServeStaticFile(assets, config, c.Request.URL.Path)
+		response := goflux.ServeStaticFile(assets, config, c.Request.URL.Path)
 
 		if response.NotFound {
 			c.AbortWithStatus(404)
