+++ adapters/nethttp/static.go
@@ -4,14 +4,14 @@ import (
 	"embed"
 	"net/http"
 
-	"github.com/barisgit/goflux/base"
+	"github.com/barisgit/goflux/goflux"
 )
 
 // StaticHandler creates a net/http handler using the shared static logic
 // Compatible with standard library mux, gorilla mux, and fasthttp (via adapter)
-func StaticHandler(assets embed.FS, config base.StaticConfig) http.Handler {
+func StaticHandler(assets embed.FS, config goflux.StaticConfig) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		response := base.ServeStaticFile(assets, config, r.URL.Path)
+		response := goflux.ServeStaticFile(assets, config, r.URL.Path)
 
 		if response.NotFound {
 			http.NotFound(w, r)
@@ -27,9 +27,9 @@ func StaticHandler(assets embed.FS, config base.StaticConfig) http.Handler {
 
 // StaticHandlerFunc creates a net/http HandlerFunc using the shared static logic
 // Alternative function signature for cases where HandlerFunc is preferred
-func StaticHandlerFunc(assets embed.FS, config base.StaticConfig) http.HandlerFunc {
+func StaticHandlerFunc(assets embed.FS, config goflux.StaticConfig) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		response := base.ServeStaticFile(assets, config, r.URL.Path)
+		response := goflux.ServeStaticFile(assets, config, r.URL.Path)
 
 		if response.NotFound {
 			http.NotFound(w, r)
