+++ goflux.go
@@ -3,7 +3,7 @@
 package goflux
 
 import (
-	"github.com/barisgit/goflux/base"
+	"github.com/barisgit/goflux/goflux"
 	"github.com/barisgit/goflux/openapi"
 
 	"github.com/danielgtaylor/huma/v2"
@@ -13,7 +13,7 @@ import (
 // AddOpenAPICommand adds an OpenAPI generation command to any cobra CLI
 // This is a convenience function that wraps the dev package
 func AddOpenAPICommand(rootCmd *cobra.Command, apiProvider func() huma.API) {
-	base.AddOpenAPICommand(rootCmd, apiProvider)
+	goflux.AddOpenAPICommand(rootCmd, apiProvider)
 }
 
 // OpenAPI generation utilities - re-export from openapi package
@@ -26,10 +26,10 @@ var (
 
 // Health check utilities - re-export from base package
 var (
-	AddHealthCheck    = base.AddHealthCheck
-	CustomHealthCheck = base.CustomHealthCheck
+	AddHealthCheck    = goflux.AddHealthCheck
+	CustomHealthCheck = goflux.CustomHealthCheck
 )
 
 // Re-export types from base package
-type StaticConfig = base.StaticConfig
-type StaticResponse = base.StaticResponse
+type StaticConfig = goflux.StaticConfig
+type StaticResponse = goflux.StaticResponse
