+++ templates/default/main.go.tmpl
@@ -61,7 +61,7 @@ import (
 	
 	// Import GoFlux framework
 	goflux "github.com/barisgit/goflux"
-	"github.com/barisgit/goflux/base"
+	"github.com/barisgit/goflux/goflux"
 )
 
 // Options for the CLI - allows configuration via flags or environment variables
@@ -252,43 +252,43 @@ func main() {
 		// Serve static files (frontend assets) - using GoFlux Chi adapter
 		// Only in production mode, development uses proxy
 		if !options.Dev {
-			router.Handle("/*", gofluxchi.StaticHandler(static.EmbeddedAssets, base.StaticConfig{SPAMode: true}))
+			router.Handle("/*", gofluxchi.StaticHandler(static.EmbeddedAssets, goflux.StaticConfig{SPAMode: true}))
 		}
 		{{else if eq .Router "fiber"}}
 		// Serve static files (frontend assets) - using GoFlux Fiber adapter
 		// Only in production mode, development uses proxy
 		if !options.Dev {
-			app.Use("/*", gofluxfiber.StaticHandler(static.EmbeddedAssets, base.StaticConfig{SPAMode: true}))
+			app.Use("/*", gofluxfiber.StaticHandler(static.EmbeddedAssets, goflux.StaticConfig{SPAMode: true}))
 		}
 		{{else if eq .Router "gin"}}
 		// Serve static files (frontend assets) - using GoFlux Gin adapter
 		// Only in production mode, development uses proxy
 		if !options.Dev {
-			router.NoRoute(gofluxgin.StaticHandler(static.EmbeddedAssets, base.StaticConfig{SPAMode: true}))
+			router.NoRoute(gofluxgin.StaticHandler(static.EmbeddedAssets, goflux.StaticConfig{SPAMode: true}))
 		}
 		{{else if eq .Router "echo"}}
 		// Serve static files (frontend assets) - using GoFlux Echo adapter
 		// Only in production mode, development uses proxy
 		if !options.Dev {
-			router.Any("/*", gofluxecho.StaticHandler(static.EmbeddedAssets, base.StaticConfig{SPAMode: true}))
+			router.Any("/*", gofluxecho.StaticHandler(static.EmbeddedAssets, goflux.StaticConfig{SPAMode: true}))
 		}
 		{{else if eq .Router "mux"}}
 		// Serve static files (frontend assets) - using GoFlux net/http adapter
 		// Only in production mode, development uses proxy
 		if !options.Dev {
-			router.Handle("/", gofluxnethttp.StaticHandler(static.EmbeddedAssets, base.StaticConfig{SPAMode: true}))
+			router.Handle("/", gofluxnethttp.StaticHandler(static.EmbeddedAssets, goflux.StaticConfig{SPAMode: true}))
 		}
 		{{else if eq .Router "gorilla"}}
 		// Serve static files (frontend assets) - using GoFlux net/http adapter
 		// Only in production mode, development uses proxy
 		if !options.Dev {
-			router.PathPrefix("/").Handler(gofluxnethttp.StaticHandler(static.EmbeddedAssets, base.StaticConfig{SPAMode: true}))
+			router.PathPrefix("/").Handler(gofluxnethttp.StaticHandler(static.EmbeddedAssets, goflux.StaticConfig{SPAMode: true}))
 		}
 		{{else if eq .Router "fasthttp"}}
 		// Serve static files (frontend assets) - using GoFlux net/http adapter
 		// Only in production mode, development uses proxy
 		if !options.Dev {
-			router.Handle("/", gofluxnethttp.StaticHandler(static.EmbeddedAssets, base.StaticConfig{SPAMode: true}))
+			router.Handle("/", gofluxnethttp.StaticHandler(static.EmbeddedAssets, goflux.StaticConfig{SPAMode: true}))
 		}
 		{{end}}
